Add files via upload
Now that I am able to add products to the database, now I need to be able to create a view that would first show the table in a separate scene. Another scene would also be able to display the Chart(s) that would be required. This will now be the majority of the project since a lot that I have done so far is not required for the project.